Creation of home page and components
The home page was created with its respective CSS design and logic, along with the importation of components. Components such as DetailCharacter, DetailLocation, and DetailEpisodes were created, each with their respective logic for communication with other components: characterList, LocationList, and EpisodeList. These components have the logic for making requests to the API created in this project.

The file RickAndMortyApi.js was created, which communicates with the API https://rickandmortyapi.com/api/, utilizing its database to develop this application.

The interface and design were developed using a combination of Tailwind and CSS. The main connector for this framework was the classes it provides, while CSS was used to keep the logic code as clean as possible. This approach creates code that is more programmer-friendly.

Functionality for data persistence was added, ensuring that data is not lost once a user logs in to the app.